+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Default the config to always set the `Accept: application/json` header https://github.com/nuwave/lighthouse/pull/743
 - Declare a single named route which handles POST/GET instead of 2 seperate routes https://github.com/nuwave/lighthouse/pull/738
 
+### Deprecated
+
+- The pagination field argument that controls the amount of results
+  will default to `first` instead of `count` in v4. The config `pagination_amount_argument`
+  can be used to change the argument name now https://github.com/nuwave/lighthouse/pull/752
+
 ## [3.4.0](https://github.com/nuwave/lighthouse/compare/v3.3.0...v3.4.0) - 2019-04-18
 
 ### Added